backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==1.10.15
SQLAlchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.6
httpx==0.27.2
passlib==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-multipart==0.0.9
pytest==8.3.3
APScheduler==3.10.4
alembic==1.13.3
pypdf==5.1.0
python-docx==1.1.2

